% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotations}
\alias{annotations}
\title{annotations: Ground-truth ratings for 100 excerpts from soundtracks}
\format{
A data frame with 100 rows and 9 variables:
\describe{
\item{nro}{File excerpt number}
\item{articulation}{Mean rating of articulation}
\item{brightness}{Mean rating of brigthness}
\item{density}{Mean rating of density}
\item{harmonic_complexity}{Mean rating of harmonic complexity}
\item{key_clarity}{Mean rating of key clarity}
\item{mode}{Mean rating of mode}
\item{pulse_clarity}{Mean rating of pulse clarity}
\item{repetitiveness}{Mean rating of repetitiveness}
...
}
}
\description{
A dataset containing ratings of 8 musical properties for 100 soundtrack excerpts.
}
