% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MusicScienceData-package.R
\docType{package}
\name{MusicScienceData-package}
\alias{MusicScienceData}
\alias{MusicScienceData-package}
\title{MusicScienceData: Example Data for Music Science tutorials}
\description{
This package holds some example datasets,
    that can be useful for teaching music and science. 
    The datasets have been derived from open accessible sources
    such as those from Harvard Dataverse or OSF.
}
\author{
\strong{Maintainer}: Tuomas Eerola \email{tuomas.eerola@durham.ac.uk} (\href{https://orcid.org/0000-0002-2896-929X}{ORCID})

}
\keyword{internal}
