% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadness.R
\name{sadness}
\alias{sadness}
\title{sadness: Attitudes towards sad music data from Eerola and Peltola 2016 study}
\format{
A data frame with 110 rows and 40 variables:
\describe{
\item{subj}{participant number}
\item{age}{Age category.}
\item{expert}{Musical expertise}
\item{gender}{Gender}
\item{listen}{Amount of music listening.}
\item{listensad}{Frequency of listening to sad music.}
\item{ASM1}{Attitude towards Sad Music Q1:}
\item{ASM2}{Attitude towards Sad Music Q1:}
\item{ASM3}{Attitude towards Sad Music Q1:}
\item{ASM4}{Attitude towards Sad Music Q1:}
\item{ASM5}{Attitude towards Sad Music Q1:}
\item{ASM6}{Attitude towards Sad Music Q1:}
\item{ASM7}{Attitude towards Sad Music Q1:}
\item{ASM8}{Attitude towards Sad Music Q1:}
\item{ASM9}{Attitude towards Sad Music Q1:}
\item{ASM10}{Attitude towards Sad Music Q1:}
\item{ASM11}{Attitude towards Sad Music Q1:}
\item{ASM12}{Attitude towards Sad Music Q1:}
\item{ASM13}{Attitude towards Sad Music Q1:}
\item{ASM14}{Attitude towards Sad Music Q1:}
\item{ASM15}{Attitude towards Sad Music Q1:}
\item{ASM16}{Attitude towards Sad Music Q1:}
\item{ASM17}{Attitude towards Sad Music Q1:}
\item{ASM18}{Attitude towards Sad Music Q1:}
\item{ASM19}{Attitude towards Sad Music Q1:}
\item{ASM20}{Attitude towards Sad Music Q1:}
\item{ASM21}{Attitude towards Sad Music Q1:}
\item{ASM22}{Attitude towards Sad Music Q1:}
\item{ASM23}{Attitude towards Sad Music Q1:}
\item{ASM24}{Attitude towards Sad Music Q1:}
\item{ASM25}{Attitude towards Sad Music Q1:}
...
}
}
\description{
A dataset containing ratings of Attitudes towards sad music from 1500+ participants.
}
