% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadness_ASM_labels.R
\name{sadness_ASM_labels}
\alias{sadness_ASM_labels}
\title{sadness Attitudes towards sad music labels}
\format{
A character string with 25 text units:
\describe{
\item{sadness_ASM_labels}{text}
...
}
}
\description{
Labels for the 25 items in the Attitudes towards Sad Music.
}
