% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundtracks.R
\name{soundtracks}
\alias{soundtracks}
\title{Soundtracks: Eerola and Vuoskoski 2011 ratings}
\format{
A data frame with 110 rows and 16 variables:
\describe{
\item{Number}{File excerpt number}
\item{Valence}{Mean rating of Valence}
\item{Energy}{Mean rating of Energy}
\item{Tension}{Mean rating of Tension}
\item{Anger}{Mean rating of Anger}
\item{Fear}{Mean rating of Fear}
\item{Happy}{Mean rating of Happy}
\item{Sad}{Mean rating of Sad}
\item{Tender}{Mean rating of Tender}
\item{Beauty}{Mean rating of Beauty}
\item{Liking}{Mean rating of Liking}
\item{TARGET_STRENGTH}{Definition of TARGET_STRENGTH (High or Moderate)}
\item{TARGET_EMOTION}{Definition of TARGET_EMOTION}
\item{TARGET_POLARITY}{Definition of TARGET_POLARITY (Neg. or Pos.)}
\item{TARGET_FRAMEWORK}{Definition of TARGET_FRAMEWORK (Discrete or Dimensional)}
\item{soundtrack}{Name of the soundtrack}
...
}
}
\description{
A dataset containing ratings of emotions for 110 film soundtrack excerpts.
}
