% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundtracks_features.R
\name{soundtracks_features}
\alias{soundtracks_features}
\title{Soundtracks features: Acoustic features of 110 soundtracks}
\format{
A data frame with 110 rows and 26 variables:
\describe{
\item{File}{File name of audio file referring to the excerpt number.}
\item{RMS}{Musical feature: RMS}
\item{sp_centr}{Spectral centroid}
\item{spec_bw}{Spectral bandwidth}
\item{spec_rolloff}{Spectral roll-off}
\item{spec_zcr}{Spectral zero-crossing}
\item{MFCC1}{Mel-cepstral coefficients nro 1}
\item{MFCC2}{Mel-cepstral coefficients nro 2}
\item{MFCC3}{Mel-cepstral coefficients nro 3}
\item{MFCC4}{Mel-cepstral coefficients nro 4}
\item{MFCC5}{Mel-cepstral coefficients nro 5}
\item{MFCC6}{Mel-cepstral coefficients nro 6}
\item{MFCC7}{Mel-cepstral coefficients nro 7}
\item{MFCC8}{Mel-cepstral coefficients nro 8}
\item{MFCC9}{Mel-cepstral coefficients nro 9}
\item{MFCC10}{Mel-cepstral coefficients nro 10}
\item{MFCC11}{Mel-cepstral coefficients nro 11}
\item{MFCC12}{Mel-cepstral coefficients nro 12}
\item{MFCC13}{Mel-cepstral coefficients nro 13}
\item{MFCC14}{Mel-cepstral coefficients nro 14}
\item{MFCC15}{Mel-cepstral coefficients nro 15}
\item{MFCC16}{Mel-cepstral coefficients nro 16}
\item{MFCC17}{Mel-cepstral coefficients nro 17}
\item{MFCC18}{Mel-cepstral coefficients nro 18}
\item{MFCC19}{Mel-cepstral coefficients nro 19}
\item{MFCC20}{Mel-cepstral coefficients nro 20}
}
}
\description{
A dataset containing acoustic features for 110 film soundtrack excerpts.
}
