% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_MusicScience.R
\name{theme_MusicScience}
\alias{theme_MusicScience}
\title{theme_MusicScience}
\usage{
theme_MusicScience(base_size = 12, base_family = "CMU Serif")
}
\arguments{
\item{base_size}{if \code{"args"} is \code{"NULL"}, base_size is one of the arguments passed to the theme function. It defaults to 12.}

\item{base_family}{if \code{"args"} is \code{"NULL"}, base_family is one of the arguments passed to the theme function. It defaults to CMU Serif.}

\item{new}{a prebuilt ggplot2 theme. Defaults to \code{"theme_minimal"}}

\item{args}{the arguments to be passed along to the ggplot2 theme function. Defaults to \code{"NULL"}.}

\item{...}{additional arguments to be used by theme}
}
\value{
None
}
\description{
This function sets a ggplot2 theme used throughout the MusicScience examples. It can be called without arguments.
}
\examples{
library(ggplot2)
qplot(hp, mpg, data=mtcars, color=am, facets=gear~cyl) + theme_MusicScience()

}
